requirements.txt
annotated-types==0.7.0
attrs==24.2.0
boto3==1.35.57
botocore==1.35.57
cachetools==5.5.2
certifi==2024.8.30
cf2tf==0.8.0
cffi==1.17.1
cfn-flip==1.3.0
charset-normalizer==3.4.0
click==8.1.7
click-log==0.4.0
configparser==7.2.0
cryptography==44.0.2
distlib==0.3.9
durationpy==0.9
filelock==3.17.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.38.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
oauthlib==3.2.2
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
rsa==4.9
s3transfer==0.10.3
setuptools==75.8.2
six==1.16.0
smmap==5.0.1
thefuzz==0.22.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.29.2
websocket-client==1.8.0
